Move creep test classes to its own module
Added MechanicalTest class and inheritance to ExperimentalStep

Added inheritance in CreepTest on MechanicalTest

Applied formatting
